Refactor finance module: migrate from Double to BigDecimal for monetary values
- Updated domain models: Transaction, Account, Budget, Goal
- Updated data layer: TransactionEntity, AccountEntity, BudgetEntity, GoalEntity
- Updated DAO: TransactionDao with BigDecimal return types
- Updated repositories: TransactionRepository, TransactionRepositoryImpl
- Updated finance utilities: MoneyFormatter, ExchangeRateProvider
- Updated ViewModels: FinanceViewModel, ProfileViewModel, TagsViewModel, TaskEditViewModel, TodayViewModel
- Updated UI components: FinanceScreen, FinanceCharts, BalanceLineChart, AddTransactionSheet
- Updated use cases: GetFinanceSummaryUseCase

This change improves precision for financial calculations and prevents floating-point rounding errors.

+++ app/src/main/java/com/taskmanager/data/local/entity/TransactionEntity.kt
@@ -3,6 +3,7 @@ package com.taskmanager.data.local.entity
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(
     tableName = "transactions",
@@ -16,7 +17,7 @@ import androidx.room.PrimaryKey
 )
 data class TransactionEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val amount: Double,
+    val amount: BigDecimal,
     val type: String,
     val currency: String = "RUB",
     val categoryId: Long? = null,
@@ -24,7 +25,7 @@ data class TransactionEntity(
     val date: Long,
     val note: String? = null,
     val toAccountId: Long? = null,
-    val destinationAmount: Double? = null,
+    val destinationAmount: BigDecimal? = null,
     val destinationCurrency: String? = null,
     val createdAt: Long = System.currentTimeMillis(),
     val updatedAt: Long = System.currentTimeMillis()
